Allow LeesEdwards solver save its state in state
LeesEdwards solver can copy its state to a struct, which can be used as
to retake it later. This is useful because many times we want to copy
the whole sytem state, including the solver. However, copying the solver
is difficult given the dealII objects it contains, which are not
copy-friendly. The alterntive is to copy the state of the solver, which
is easy, and use always the same solver object.