Add the shear and bulk modulus evolution to the macro evol. history
The values corresponds to the average and std, and not to the global
effective values (the average and the effective are the same if the
properties as spatially homogeneous, i.e. if the std=0). The values are
just read from the element's stiffness tensor, i.e. the shear modulus is
defined as the C_xyxy component and the bulk modulus the (C_xxxx+C_xxyy)/2.