Allow the global effective elastic properties evolve with av. plastic strain
We consider elastic properties that are different in the quench state and in the
stationary state. Allowing the properties to change locally after each plastic
event is computationally very expensive. Instead, we consider their global change.
For this, we consider homogenous elastic properties, and make them evolve smoothly
with the average plastic deformation. The solver is reassembled whenever the change
of the accumulated change in the shear modulus reaches a certain threshold. The
transition between quench and stationary properties occurs over a characteristic
plastic strain value, introduced as a new simulation parameter.