Compute the macro evolution summary within the base history class
The responsibility of computing the macro evolution summary is of the
base history class. Some magnitudes are removed and another ones of a
more general relevance are added (the stress tensor components). If new
magnitudes are needed, the user can derived from the base class and
extend the macro evolution summary.