Overload Element's make_copy function in derived classes
The element's derived classes overload the make_copy function of the base class,
returning a pointer to the derive class type instead of base type. In this way,
when a copy is requested from a pointer to base, we get pointer to base as the
result. However, if it is requested from a pointer to a derived class, we get a
pointer to derived class. This makes it easier to have access to the
configuration struct of the elements and does not break polymorphism.