Add Launcher class
The Launcher class allows to run simulations in parallel easily using OpenMP.
The users only need to create a class that inherits from the Launcher class and
inplement the run_impl() function. The simulation is the executed calling run().
If the code was compiled using OpenMP, it will run in parallel. If not, it will
default to a single-threaded serial run.