Make the slip properties vary continuously with the local plastic strain
Before, the slip properties were updated according to the local von Mises plastic
strain but only after each avalanche. Now, the elements know the values of the
quench and the renewed properties and perform the variation in the
renew_thresholds() function, that is, as continously as possible with the change
in local plastic deformation.