Stop converting quench elastic states into a pre-stress
When the quench elastic state is created, it is left as an eigenstrain field in
the solver (and only in the solver). The elements are not added eigenstrain, or
it is cleaned before doing AQS. Not "freezing" the stress into the presterss
member ensures that if the elastic properties are varied with the global strain,
the part of the elastic state coming from the quench state, will also be
affected (imagine e.g. that there's an intense quench stress field, and the system
is heated up and the elastic properties become very low, in such case that stress
field should drop accordingly with the change in the elastic properties).