Update plastic event after calling add() with a single event
When calling the overloaded add() with a single event, a vector with a copy of
it is created and passed to the main add() function. Therefore, the user cannot
see the updated info added by this function call to the event. To solve this, the
updated event is copied back into the input one. This operation does not add
overhead since the add() function also involves solving the FEM system,
which is many orders of magnitude more expensive.