Remove the namespace "quench" and reorganize the simulation structure
We refer to quench to the operation by which the liquid solidifies. Thus, the
creation of the prestress field using an eigenstrain field should not be part
of that operation, but of the simulation general setup. The quench section is
now renamed as "simulate parent liquid". Since prestress is an eigenstrain field,
the function equilibrate_initial_structure_relaxation has been removed because
is obsolete.