Add elastic and configurational energy members to the base element class
The elastic energy is computed from the stress and the stiffness tensor using
linear elasticity. The configurational energy is an extra energy which must be
added by the user, since its origins are not elastic. MEPLS now queries the
elements for their energy instead of computing it at different places.