Replace the element copy constructor with a method that copies the state
Instead of implementing the copy constructor, we define a function that copies
member-wise the entire state of an element into another.

The reason is that sometimes, we want to revert an element's state (think of
Metropolis-Hastings algorithm when a candidate event is rejected after having
been performed). To revert the state, we would make a copy-element before the
change and then change the pointer in the elements vector to the copied element
and delete the original. This has the inconvenience of invalidating the pointers
in other possible copies of the vector of (pointers to) elements that might exist
in other places. To avoid invalidating the pointers, we do not delete the
original elements. Instead, we modify their state by copying member-wise the
state of the untouched copied element. The copied element can then be safely
deleted.