Update CMakeLists
cmake now requires an enviroment variable MEPLS_DIR. Before,
the tutorials and the gallery models were using relative
paths. This works for the MEPLS bundled models, but it is
better to provide the users with CMakeLists examples that
are not dependent on relative paths and that would work
wherever they place the source code.

cmake now forces that deal.II version is 9.0, since other
versions (at least much older or much newer) won't compile.